requirements.txt
Flask==2.3.3
numpy
Pillow
requests
matplotlib
h5py
rasterio
python-dotenv
